Allow LLVM compiler to be correctly detected and used on MacOSX
(1) Because LLVM brings along its own ar/runlib replacements, they have to be enforced
to avoid collision with the potentially incompatible default versions.

(2) LLVM has switched to using libstdc++ instead of libc++ by default, but for compatibility
with objects compiled separately we need to restore the previous behaviour.

(3) Several shell scripts need to be passed .dylib as the filename suffix for dynamically-linked libraries.
An existing variable CK_DLL_EXT is (re-)used for this purpose.