bigfix: make sure ICC compiler sets all the necessary paths to its tools for cmake to operate correctly
Here we are passing the paths to ICC's equivalent to LD (xild), AR (xiar) and RANLIB (/usr/bin/ranlib)
to cmake via CK_..._PATH_FOR_CMAKE environment variables.

ICC doesn't have a separate utility for ranlib, and uses "xiar s" instead.
However, the current version of cmake does not support parameters to ranlib.
Passing the combined "/path/to/xiar s" string to cmake (as a substitute for ranlib)
proved to be very difficult, despite all kinds of escaping.

A better solution would be for the soft entry to create an executable ranlib_wrapper
which would reside in the generated env entry, and the CK_RANLIB_PATH_FOR_CMAKE would point to that.
Unfortunately, at the moment the structure of the code is not very conductive to this approach.
So we are staying with the current non-elegant solution for the time being.