Improve consistency off remote_shell start and end
CPU and GPU frequency setting only called remote_shell and not remote_shell_end.
This left some commands unquoted if quotes were present in both remote_shell and
remote_shell_end. Implementation is not as tidy as other cmd calls as I wanted
to preserve backwards compatibility. If not, all targets would require
remote_shell and remote_shell_end to contain quotes to ensure cmds were
correctly executed on the remote targets.

The situation that I found myself in was:
+ Using a linux remote target with access type ssh
+ The detect function in platform.os (ck-env/module/platform.os/module.py)
  performs a chmod where the string required remote_shell and remote_shell_end
  to have quotes. Otherwise the command was executed incorrectly
+ Adding quotes to remote_shell and remote_shell_end fixed the above issue but
  platform.cpu and platform.gpu set_freq functions created a cmd which had
  three quotes producing and error.